Update concurrency for updatecontent
We can't use a Go routine for each file because swupd.Hashcalc() calls
os.Open() on each file passed to it, which can quickly overflow the max
FD limit. This patch clamps it to the number of CPUs * 3, or the number
of files to process, whichever is smaller. Further testing should be done
if this shows to be a bottleneck.

Fixes issue #82

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>